Bug 4888 - pre-set of field indicators value
Preload on database of the set of field indicators value as defined by
http://www.loc.gov/marc/bibliographic/ecbdhome.html.
New sql files in marc21/mandatory to create and populate the indicators tables with the sets of values for
the Default framework/authtype.
Each framework/authtype will be able to have its own copy to allow individual modification of indicators value.
When editing the framework/authtype exists the possibility of cloning the indicators set from an existing
framework/authtype populated with indicators.

Modification of the intranet administration-Marc Frameworks/Authority MARC Framework to clone on creation from
another framework and to delete set of indicators from a framework.
Management (add/modify/delete) for indicators of the fields in the Framework Structure.

Modification of the intranet cataloguing addbiblio / authorities authorities to enable a plugin to see the
defined indicators value for a specific framework, and to check validity of indicators value.

Two new system preferences: CheckValueIndicators for checking the value of the indicators when cataloguing
a biblio record; DisplayPluginValueIndicators: to enable the plugin at addbiblio to see indicators value and to
check too.