Local repository
Local repository of cover images and documents related with the biblio and documents related with the
items.
In the cataloguing section when adding a biblio exists the possibility of attach an image and/or a document
to the biblio. The files uploaded are saved in the path specified in the system preference "dirFileLocalRepository".
The names of the files are  "biblionumber.extension" uploaded.
The 856$u is changed to store the new url of the document, and the 856$y has "See document" as value.
When adding an item exists the possibility of attach a document to the item.
The names of the files uploaded are "biblionumber_itemnumber.extension".

Three new system preferences:
"dirFileLocalRepository": stores the path of the local repository.
"dirUrlLocalRepository": stores the url to access from http portocol of the local repository.
"LocalRepositoryCoverImages": show cover images if available.
"MaxSizeFileLocalRepository": max size to uploaded file in the local repository.